A data mining project with using R
Glioblastoma Multiforme, also known as GBM, GBM-IV or grade 4 tumor, is the most common type of malignant brain tumor. Glioblastoma consists of astrocytomas, the highest grade (grade 4) type of gliomas. Gliomas are groups of tumors that form in glia cells that protect and support nerve cells. GBM is a primary malignant brain tumor and it does not have a definite treatment besides its rapid progression. The project goal is to search and analyze gene activity in Glioblastoma Multiforme patients from the literature, the web and datasets.